Increased accuracy

+++ breakPrediction/server.py
@@ -21,6 +21,6 @@ def predict_endpoint(request: PredictRequest):
             period="Day"
 
         prediction1 = prediction.predict_breaks(period, request.totalHours, request.age)
-        return {"break": prediction1*request.totalHours,"period":period}
+        return {"break": int(prediction1*request.totalHours),"period":period}
     except Exception as e:
         raise HTTPException(status_code=400, detail=str(e))
